requirements.txt
Pillow==9.4.0
streamlit==1.11.0
streamlit_player==0.1.5
